Refactored structure converter
