+++ .idea/gradle.xml
@@ -8,7 +8,6 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/StructureConverter" />
             <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
