--- StructureConverter/build.gradle
@@ -1,8 +1,0 @@
-apply plugin: 'java-library'
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-}
-
-sourceCompatibility = "8"
-targetCompatibility = "8"
