+++ app/src/main/java/c/example/speechtobsl/outer_framework/SpeechInputActivity.java
@@ -1,28 +1,19 @@
 package c.example.speechtobsl.outer_framework;
 
 import android.Manifest;
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.content.pm.PackageManager;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.SystemClock;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import c.example.speechtobsl.R;
-import c.example.speechtobsl.models.StructureConverterModel;
 import c.example.speechtobsl.views.SpeechView;
 
 
@@ -135,9 +126,6 @@ public class SpeechInputActivity extends AppCompatActivity implements SuccessLis
         @Override
         protected void onCancelled() {
             super.onCancelled();
-            mRecordButton.setVisibility(View.VISIBLE);
-            mRecordText.setVisibility(View.VISIBLE);
-            mLoadingText.setVisibility(View.INVISIBLE);
         }
 
         @Override
