+++ app/src/main/java/c/example/speechtobsl/structure_converter/utils/POS.java
@@ -1,0 +1,5 @@
+package c.example.speechtobsl.structure_converter.utils;
+
+public enum POS {
+        NOUN, VRB, DET, ADJ, ADV, CONN, TIM, MDL, QTN, PREP, NA
+}
