+++ app/src/main/res/layout/activity_main.xml
@@ -22,16 +22,50 @@
 
     <include layout="@layout/content_main" />
 
+    <RelativeLayout
+        android:id="@+id/topLevel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="100dp">
+
+    <TextView
+        android:id="@+id/title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:text="Speech To BSL"
+        android:textSize="36sp" />
+
+    </RelativeLayout>
+
+    <RelativeLayout
+        android:id="@+id/middleLevel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:layout_centerVertical="true"
+        android:layout_marginBottom="10dp">
+
+    <TextView
+        android:id="@+id/loading_text"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:text="TextView"
+        android:textSize="24sp"
+        tools:text="Loading..." />
     <TextView
         android:id="@+id/record_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="300dp"
-        android:layout_marginBottom="-300dp"
         android:text="TextView"
         android:textSize="18sp"
         tools:text="Press the red button to start recording" />
+
+    </RelativeLayout>
 
     <RelativeLayout
         android:id="@+id/bottomLevel"
@@ -43,33 +77,12 @@
 
         <Button
             android:id="@+id/record_button"
-
             android:layout_width="70dp"
             android:layout_height="70dp"
             android:background="@drawable/not_recording"/>
 
 
     </RelativeLayout>
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="134dp"
-        android:text="Speech To BSL"
-        android:textSize="36sp" />
-
-    <TextView
-        android:id="@+id/loading_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="300dp"
-        android:text="TextView"
-        android:textSize="24sp"
-        tools:text="Loading..." />
 
 
 </RelativeLayout>
