+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':StructureConverter'
+include ':app'
